requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
sqlalchemy==2.0.23
pymysql==1.1.1
python-dotenv==1.0.0
python-jose[cryptography]==3.4.0
passlib==1.7.4
bcrypt==4.0.1
alembic==1.12.0
pydantic~=2.10.6
pdfplumber~=0.11.5
python-docx~=1.1.2
pytesseract~=0.3.13
pillow~=11.1.0
ollama~=0.4.7
python-multipart
pydantic[email]